refactor: extract Domain migration rewrite helpers under 500 lines
Extract Class for the imperative Domain migration orchestrators: move the
legacy-assembly migration context/capabilities out of CSharpRules, CaptureWindow
rewrites out of ScreenshotRules, and application/first-party type rewrites out
of TypeReplacementRules. No behavior change; characterization and RulesTests stay green.
