+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationApplicationAndFirstPartyTypeReplacementRules.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 88ffdd63a32a49b28ee09f444f6c1b67
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
