+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationCSharpLegacyAssemblyMigrationCapabilities.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 97896d4b136343feab60d74f8c75e99a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
