+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationCSharpLegacyAssemblyMigrationContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff7fe0fd41dd4a19912624c743aba1e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
