+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationCSharpRules.cs
@@ -45,6 +45,7 @@ using static io.github.hatayama.UnityCliLoop.Domain.ThirdPartyToolMigrationTimin
 using static io.github.hatayama.UnityCliLoop.Domain.ThirdPartyToolMigrationTimingTypeScopeRules;
 using static io.github.hatayama.UnityCliLoop.Domain.ThirdPartyToolMigrationToolContractDetectionRules;
 using static io.github.hatayama.UnityCliLoop.Domain.ThirdPartyToolMigrationTypeReplacementRules;
+using static io.github.hatayama.UnityCliLoop.Domain.ThirdPartyToolMigrationApplicationAndFirstPartyTypeReplacementRules;
 
 namespace io.github.hatayama.UnityCliLoop.Domain
 {
@@ -119,639 +120,6 @@ namespace io.github.hatayama.UnityCliLoop.Domain
             migrationContext.ApplyRegistrarRenames();
             migrationContext.ApplyCurrentPublicContractNamespaceReplacements();
             return migrationContext.CreateResult();
-        }
-
-        /// <summary>
-        /// Carries the derived migration context while applying C# source rewrite phases.
-        /// </summary>
-        private sealed class CSharpLegacyAssemblyMigrationContext
-        {
-            private readonly string[] _legacyNamespaceAliases;
-            private readonly string[] _legacyAssemblyToolInfoAliases;
-            private readonly string[] _currentApplicationNamespaceAliases;
-            private readonly string[] _currentDomainNamespaceAliases;
-            private readonly string[] _currentFirstPartyToolsNamespaceAliases;
-            private readonly string[] _assemblyDeclaredTypeNames;
-            private readonly bool _canUseCurrentToolContracts;
-            private readonly bool _canUseBareCurrentToolContracts;
-            private readonly bool _canPreserveBareCurrentToolContractsReferences;
-            private readonly bool _canMigrateBareLegacyToolAttribute;
-            private readonly bool _canMigrateBareLegacyEditorWindowCaptureUtility;
-            private readonly bool _canUseBareCurrentFirstPartyTools;
-            private readonly bool _shouldQualifyBareEditorWindowCaptureUtilityTimeout;
-            private readonly bool _canMigrateBareLegacyFirstPartyScreenshotApi;
-            private readonly bool _canMigrateBareCurrentDomainContractType;
-            private readonly bool _canMigrateBareLegacyApplicationApi;
-            private readonly bool _canMigrateBareLegacyApplicationTypeName;
-            private readonly bool _canMigrateAmbiguousBareLegacyToolInfoConstructor;
-            private readonly bool _shouldApplyContractRenames;
-            private readonly bool _shouldApplyRegistrarRenames;
-            private readonly bool _shouldApplyDomainMetadataRenames;
-            private readonly List<RemovedLegacyPlayerLoopTimingSignature> _removedPlayerLoopTimingSignatures =
-                new();
-            private string _migratedContent;
-            private int _replacementCount;
-
-            private CSharpLegacyAssemblyMigrationContext(
-                string source,
-                string[] legacyNamespaceAliases,
-                string[] legacyAssemblyToolInfoAliases,
-                string[] currentApplicationNamespaceAliases,
-                string[] currentDomainNamespaceAliases,
-                string[] currentFirstPartyToolsNamespaceAliases,
-                string[] assemblyDeclaredTypeNames,
-                bool canUseCurrentToolContracts,
-                bool canUseBareCurrentToolContracts,
-                bool canPreserveBareCurrentToolContractsReferences,
-                bool canMigrateBareLegacyToolAttribute,
-                bool canMigrateBareLegacyEditorWindowCaptureUtility,
-                bool canUseBareCurrentFirstPartyTools,
-                bool shouldQualifyBareEditorWindowCaptureUtilityTimeout,
-                bool canMigrateBareLegacyFirstPartyScreenshotApi,
-                bool canMigrateBareCurrentDomainContractType,
-                bool canMigrateBareLegacyApplicationApi,
-                bool canMigrateBareLegacyApplicationTypeName,
-                bool canMigrateAmbiguousBareLegacyToolInfoConstructor,
-                bool shouldApplyContractRenames,
-                bool shouldApplyRegistrarRenames,
-                bool shouldApplyDomainMetadataRenames)
-            {
-                _migratedContent = source;
-                _legacyNamespaceAliases = legacyNamespaceAliases;
-                _legacyAssemblyToolInfoAliases = legacyAssemblyToolInfoAliases;
-                _currentApplicationNamespaceAliases = currentApplicationNamespaceAliases;
-                _currentDomainNamespaceAliases = currentDomainNamespaceAliases;
-                _currentFirstPartyToolsNamespaceAliases = currentFirstPartyToolsNamespaceAliases;
-                _assemblyDeclaredTypeNames = assemblyDeclaredTypeNames;
-                _canUseCurrentToolContracts = canUseCurrentToolContracts;
-                _canUseBareCurrentToolContracts = canUseBareCurrentToolContracts;
-                _canPreserveBareCurrentToolContractsReferences = canPreserveBareCurrentToolContractsReferences;
-                _canMigrateBareLegacyToolAttribute = canMigrateBareLegacyToolAttribute;
-                _canMigrateBareLegacyEditorWindowCaptureUtility =
-                    canMigrateBareLegacyEditorWindowCaptureUtility;
-                _canUseBareCurrentFirstPartyTools = canUseBareCurrentFirstPartyTools;
-                _shouldQualifyBareEditorWindowCaptureUtilityTimeout =
-                    shouldQualifyBareEditorWindowCaptureUtilityTimeout;
-                _canMigrateBareLegacyFirstPartyScreenshotApi = canMigrateBareLegacyFirstPartyScreenshotApi;
-                _canMigrateBareCurrentDomainContractType = canMigrateBareCurrentDomainContractType;
-                _canMigrateBareLegacyApplicationApi = canMigrateBareLegacyApplicationApi;
-                _canMigrateBareLegacyApplicationTypeName = canMigrateBareLegacyApplicationTypeName;
-                _canMigrateAmbiguousBareLegacyToolInfoConstructor =
-                    canMigrateAmbiguousBareLegacyToolInfoConstructor;
-                _shouldApplyContractRenames = shouldApplyContractRenames;
-                _shouldApplyRegistrarRenames = shouldApplyRegistrarRenames;
-                _shouldApplyDomainMetadataRenames = shouldApplyDomainMetadataRenames;
-            }
-
-            public static CSharpLegacyAssemblyMigrationContext Create(
-                string source,
-                bool hasLegacyAssemblySource,
-                bool hasAssemblyScopedCurrentToolContractsUsing,
-                bool hasAssemblyScopedCurrentApplicationUsing,
-                bool hasAssemblyScopedCurrentDomainUsing,
-                bool hasAssemblyScopedCurrentFirstPartyToolsUsing,
-                string[] legacyAssemblyAliases,
-                string[] legacyAssemblyToolInfoAliases,
-                string[] currentApplicationAssemblyAliases,
-                string[] currentDomainAssemblyAliases,
-                string[] currentFirstPartyToolsAssemblyAliases,
-                string[] assemblyDeclaredTypeNames)
-            {
-                string[] legacyNamespaceAliases = GetCombinedLegacyNamespaceAliases(source, legacyAssemblyAliases);
-                string[] currentApplicationNamespaceAliases = GetCombinedCurrentApplicationNamespaceAliases(
-                    source,
-                    currentApplicationAssemblyAliases);
-                string[] currentDomainNamespaceAliases = GetCombinedCurrentDomainNamespaceAliases(
-                    source,
-                    currentDomainAssemblyAliases);
-                string[] currentFirstPartyToolsNamespaceAliases =
-                    GetCombinedCurrentFirstPartyToolsNamespaceAliases(
-                        source,
-                        currentFirstPartyToolsAssemblyAliases);
-                CSharpLegacyAssemblyMigrationCapabilities capabilities =
-                    CSharpLegacyAssemblyMigrationCapabilities.Create(
-                        source,
-                        hasLegacyAssemblySource,
-                        hasAssemblyScopedCurrentToolContractsUsing,
-                        hasAssemblyScopedCurrentApplicationUsing,
-                        hasAssemblyScopedCurrentDomainUsing,
-                        hasAssemblyScopedCurrentFirstPartyToolsUsing,
-                        legacyNamespaceAliases);
-                return new CSharpLegacyAssemblyMigrationContext(
-                    source,
-                    legacyNamespaceAliases,
-                    legacyAssemblyToolInfoAliases,
-                    currentApplicationNamespaceAliases,
-                    currentDomainNamespaceAliases,
-                    currentFirstPartyToolsNamespaceAliases,
-                    assemblyDeclaredTypeNames,
-                    capabilities.CanUseCurrentToolContracts,
-                    capabilities.CanUseBareCurrentToolContracts,
-                    capabilities.CanPreserveBareCurrentToolContractsReferences,
-                    capabilities.CanMigrateBareLegacyToolAttribute,
-                    capabilities.CanMigrateBareLegacyEditorWindowCaptureUtility,
-                    capabilities.CanUseBareCurrentFirstPartyTools,
-                    capabilities.ShouldQualifyBareEditorWindowCaptureUtilityTimeout,
-                    capabilities.CanMigrateBareLegacyFirstPartyScreenshotApi,
-                    capabilities.CanMigrateBareCurrentDomainContractType,
-                    capabilities.CanMigrateBareLegacyApplicationApi,
-                    capabilities.CanMigrateBareLegacyApplicationTypeName,
-                    capabilities.CanMigrateAmbiguousBareLegacyToolInfoConstructor,
-                    capabilities.ShouldApplyContractRenames,
-                    capabilities.ShouldApplyRegistrarRenames,
-                    capabilities.ShouldApplyDomainMetadataRenames);
-            }
-
-            public void ApplyToolMetadataReplacements()
-            {
-                _migratedContent = ReplaceLegacyToolAttributesInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    _canMigrateBareLegacyToolAttribute,
-                    ref _replacementCount);
-                _migratedContent = ReplaceLegacyToolInfoConstructorsInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    _canMigrateBareLegacyToolAttribute,
-                    _canMigrateAmbiguousBareLegacyToolInfoConstructor,
-                    _legacyAssemblyToolInfoAliases,
-                    ref _replacementCount);
-                _migratedContent = ReplaceLegacyToolSettingsCatalogItemConstructorsInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    _canMigrateBareLegacyToolAttribute,
-                    ref _replacementCount);
-            }
-
-            public void ApplyDelayReplacements()
-            {
-                (string editorDelayMigratedContent, int editorDelayReplacementCount) =
-                    ReplaceLegacyEditorDelayFrameCallsInCode(
-                        _migratedContent,
-                        _legacyNamespaceAliases,
-                        _canMigrateBareLegacyToolAttribute || _canUseCurrentToolContracts,
-                        !_canUseBareCurrentToolContracts);
-                ApplyReplacementResult(editorDelayMigratedContent, editorDelayReplacementCount);
-                (string timerDelayMigratedContent, int timerDelayReplacementCount) =
-                    ReplaceLegacyTimerDelayNamedArgumentsInCode(
-                        _migratedContent,
-                        _legacyNamespaceAliases,
-                        _canMigrateBareLegacyToolAttribute || _canUseCurrentToolContracts);
-                ApplyReplacementResult(timerDelayMigratedContent, timerDelayReplacementCount);
-            }
-
-            public void ApplyMainThreadSwitcherReplacements()
-            {
-                (string migratedContent, int replacementCount) =
-                    ReplaceLegacyMainThreadSwitcherCallsInCode(
-                        _migratedContent,
-                        _legacyNamespaceAliases,
-                        _currentApplicationNamespaceAliases,
-                        _canMigrateBareLegacyApplicationApi,
-                        _assemblyDeclaredTypeNames);
-                ApplyReplacementResult(migratedContent, replacementCount);
-                if (replacementCount == 0 && !ContainsMigratedMainThreadSwitcherSwitchCall(_migratedContent))
-                {
-                    return;
-                }
-
-                RemovePlayerLoopTimingParameters();
-                RemoveUnusedPlayerLoopTimingDeclarations();
-            }
-
-            public void ApplyEditorWindowCaptureAndTypeReplacements()
-            {
-                (string editorWindowCaptureMigratedContent, int editorWindowCaptureReplacementCount) =
-                    ReplaceLegacyEditorWindowCaptureUtilityCallsInCode(
-                        _migratedContent,
-                        _legacyNamespaceAliases,
-                        _currentFirstPartyToolsNamespaceAliases,
-                        _canMigrateBareLegacyEditorWindowCaptureUtility,
-                        _shouldQualifyBareEditorWindowCaptureUtilityTimeout,
-                        _canPreserveBareCurrentToolContractsReferences,
-                        _canUseBareCurrentFirstPartyTools,
-                        _assemblyDeclaredTypeNames);
-                ApplyReplacementResult(editorWindowCaptureMigratedContent, editorWindowCaptureReplacementCount);
-                _migratedContent = ReplaceLegacyFirstPartyScreenshotTypeNamesInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    _currentFirstPartyToolsNamespaceAliases,
-                    _canMigrateBareLegacyFirstPartyScreenshotApi,
-                    _canPreserveBareCurrentToolContractsReferences,
-                    _assemblyDeclaredTypeNames,
-                    ref _replacementCount);
-                _migratedContent = ReplaceCurrentDomainContractTypeNamesInCode(
-                    _migratedContent,
-                    _currentDomainNamespaceAliases,
-                    _canMigrateBareCurrentDomainContractType,
-                    _canPreserveBareCurrentToolContractsReferences,
-                    _assemblyDeclaredTypeNames,
-                    ref _replacementCount);
-                _migratedContent = ReplaceLegacyApplicationTypeNamesInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    _currentApplicationNamespaceAliases,
-                    _canMigrateBareLegacyApplicationTypeName,
-                    _canPreserveBareCurrentToolContractsReferences,
-                    _assemblyDeclaredTypeNames,
-                    ref _replacementCount);
-                _migratedContent = ReplaceLegacyRegistrarAliasesInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    ref _replacementCount);
-            }
-
-            public void ApplyContractRenames()
-            {
-                if (!_shouldApplyContractRenames)
-                {
-                    return;
-                }
-
-                _migratedContent = ReplaceLegacyDomainTypeNamesInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    _assemblyDeclaredTypeNames,
-                    ref _replacementCount);
-                _migratedContent = ReplaceLegacyContractTypeNamesInCode(
-                    _migratedContent,
-                    _legacyNamespaceAliases,
-                    _assemblyDeclaredTypeNames,
-                    ref _replacementCount);
-                ApplyCSharpReplacementRules();
-            }
-
-            public void ApplyRegistrarRenames()
-            {
-                if (!_shouldApplyRegistrarRenames && !_shouldApplyDomainMetadataRenames)
-                {
-                    return;
-                }
-
-                if (_shouldApplyRegistrarRenames)
-                {
-                    _migratedContent = ReplaceUnqualifiedLegacyRegistrarReferencesInCode(
-                        _migratedContent,
-                        ref _replacementCount);
-                }
-
-                ApplyRegistrarReplacementRules();
-                _migratedContent = ReplaceLegacyToolInfoTypeReferencesInCode(
-                    _migratedContent,
-                    _assemblyDeclaredTypeNames,
-                    ref _replacementCount);
-            }
-
-            public void ApplyCurrentPublicContractNamespaceReplacements()
-            {
-                _migratedContent = ReplaceCurrentPublicContractNamespacesInCode(
-                    _migratedContent,
-                    ref _replacementCount);
-            }
-
-            public ThirdPartyToolMigrationContentResult CreateResult()
-            {
-                return new ThirdPartyToolMigrationContentResult(
-                    _migratedContent,
-                    _replacementCount,
-                    _removedPlayerLoopTimingSignatures.ToArray());
-            }
-
-            private void RemovePlayerLoopTimingParameters()
-            {
-                string[] migratedCalleeMethodNames = Array.Empty<string>();
-                while (true)
-                {
-                    (
-                        string migratedContent,
-                        int replacementCount,
-                        RemovedLegacyPlayerLoopTimingSignature[] localRemovedTimingSignatures) =
-                        RemoveLegacyPlayerLoopTimingParametersInCode(
-                            _migratedContent,
-                            _legacyNamespaceAliases,
-                            _canMigrateBareLegacyApplicationApi,
-                            migratedCalleeMethodNames);
-                    if (replacementCount == 0)
-                    {
-                        return;
-                    }
-
-                    ApplyReplacementResult(migratedContent, replacementCount);
-                    _removedPlayerLoopTimingSignatures.AddRange(localRemovedTimingSignatures);
-                    (string timingCallerMigratedContent, int timingCallerReplacementCount) =
-                        RemoveLegacyPlayerLoopTimingCallerArgumentsInCode(
-                            _migratedContent,
-                            localRemovedTimingSignatures,
-                            _legacyNamespaceAliases);
-                    ApplyReplacementResult(timingCallerMigratedContent, timingCallerReplacementCount);
-                    migratedCalleeMethodNames = localRemovedTimingSignatures
-                        .Select(signature => signature.MethodName)
-                        .ToArray();
-                }
-            }
-
-            private void RemoveUnusedPlayerLoopTimingDeclarations()
-            {
-                (string migratedContent, int replacementCount) =
-                    RemoveUnusedLegacyPlayerLoopTimingDeclarationsInCode(_migratedContent);
-                ApplyReplacementResult(migratedContent, replacementCount);
-            }
-
-            private void ApplyCSharpReplacementRules()
-            {
-                foreach (ReplacementRule rule in CSharpReplacementRules)
-                {
-                    _migratedContent = ReplaceRegexInCode(
-                        _migratedContent,
-                        rule.PatternRegex,
-                        _ => rule.Replacement,
-                        ref _replacementCount);
-                }
-            }
-
-            private void ApplyRegistrarReplacementRules()
-            {
-                foreach (ReplacementRule rule in RegistrarReplacementRules)
-                {
-                    _migratedContent = ReplaceRegexInCode(
-                        _migratedContent,
-                        rule.PatternRegex,
-                        _ => rule.Replacement,
-                        ref _replacementCount);
-                }
-            }
-
-            private void ApplyReplacementResult(string migratedContent, int replacementCount)
-            {
-                _migratedContent = migratedContent;
-                _replacementCount += replacementCount;
-            }
-        }
-
-        private readonly struct CSharpLegacyAssemblyMigrationCapabilities
-        {
-            private CSharpLegacyAssemblyMigrationCapabilities(
-                bool canUseCurrentToolContracts,
-                bool canUseBareCurrentToolContracts,
-                bool canPreserveBareCurrentToolContractsReferences,
-                bool canMigrateBareLegacyToolAttribute,
-                bool canMigrateBareLegacyEditorWindowCaptureUtility,
-                bool canUseBareCurrentFirstPartyTools,
-                bool shouldQualifyBareEditorWindowCaptureUtilityTimeout,
-                bool canMigrateBareLegacyFirstPartyScreenshotApi,
-                bool canMigrateBareCurrentDomainContractType,
-                bool canMigrateBareLegacyApplicationApi,
-                bool canMigrateBareLegacyApplicationTypeName,
-                bool canMigrateAmbiguousBareLegacyToolInfoConstructor,
-                bool shouldApplyContractRenames,
-                bool shouldApplyRegistrarRenames,
-                bool shouldApplyDomainMetadataRenames)
-            {
-                CanUseCurrentToolContracts = canUseCurrentToolContracts;
-                CanUseBareCurrentToolContracts = canUseBareCurrentToolContracts;
-                CanPreserveBareCurrentToolContractsReferences = canPreserveBareCurrentToolContractsReferences;
-                CanMigrateBareLegacyToolAttribute = canMigrateBareLegacyToolAttribute;
-                CanMigrateBareLegacyEditorWindowCaptureUtility = canMigrateBareLegacyEditorWindowCaptureUtility;
-                CanUseBareCurrentFirstPartyTools = canUseBareCurrentFirstPartyTools;
-                ShouldQualifyBareEditorWindowCaptureUtilityTimeout =
-                    shouldQualifyBareEditorWindowCaptureUtilityTimeout;
-                CanMigrateBareLegacyFirstPartyScreenshotApi = canMigrateBareLegacyFirstPartyScreenshotApi;
-                CanMigrateBareCurrentDomainContractType = canMigrateBareCurrentDomainContractType;
-                CanMigrateBareLegacyApplicationApi = canMigrateBareLegacyApplicationApi;
-                CanMigrateBareLegacyApplicationTypeName = canMigrateBareLegacyApplicationTypeName;
-                CanMigrateAmbiguousBareLegacyToolInfoConstructor =
-                    canMigrateAmbiguousBareLegacyToolInfoConstructor;
-                ShouldApplyContractRenames = shouldApplyContractRenames;
-                ShouldApplyRegistrarRenames = shouldApplyRegistrarRenames;
-                ShouldApplyDomainMetadataRenames = shouldApplyDomainMetadataRenames;
-            }
-
-            public bool CanUseCurrentToolContracts { get; }
-            public bool CanUseBareCurrentToolContracts { get; }
-            public bool CanPreserveBareCurrentToolContractsReferences { get; }
-            public bool CanMigrateBareLegacyToolAttribute { get; }
-            public bool CanMigrateBareLegacyEditorWindowCaptureUtility { get; }
-            public bool CanUseBareCurrentFirstPartyTools { get; }
-            public bool ShouldQualifyBareEditorWindowCaptureUtilityTimeout { get; }
-            public bool CanMigrateBareLegacyFirstPartyScreenshotApi { get; }
-            public bool CanMigrateBareCurrentDomainContractType { get; }
-            public bool CanMigrateBareLegacyApplicationApi { get; }
-            public bool CanMigrateBareLegacyApplicationTypeName { get; }
-            public bool CanMigrateAmbiguousBareLegacyToolInfoConstructor { get; }
-            public bool ShouldApplyContractRenames { get; }
-            public bool ShouldApplyRegistrarRenames { get; }
-            public bool ShouldApplyDomainMetadataRenames { get; }
-
-            public static CSharpLegacyAssemblyMigrationCapabilities Create(
-                string source,
-                bool hasLegacyAssemblySource,
-                bool hasAssemblyScopedCurrentToolContractsUsing,
-                bool hasAssemblyScopedCurrentApplicationUsing,
-                bool hasAssemblyScopedCurrentDomainUsing,
-                bool hasAssemblyScopedCurrentFirstPartyToolsUsing,
-                string[] legacyNamespaceAliases)
-            {
-                CSharpMigrationSourceUsage usage = DetectMigrationSourceUsage(source);
-                bool canUseCurrentToolContracts = CanUseCurrentToolContractsFromUsage(
-                    usage,
-                    hasAssemblyScopedCurrentToolContractsUsing);
-                bool canUseBareCurrentToolContracts = CanUseBareCurrentToolContractsFromUsage(
-                    usage,
-                    hasLegacyAssemblySource,
-                    hasAssemblyScopedCurrentToolContractsUsing);
-                bool canMigrateBareLegacyToolAttribute = CanMigrateBareLegacyToolAttributeFromUsage(
-                    usage,
-                    hasLegacyAssemblySource,
-                    legacyNamespaceAliases);
-                bool canMigrateBareLegacyApplicationApi = CanMigrateBareLegacyApplicationApiFromUsage(
-                    usage,
-                    canMigrateBareLegacyToolAttribute,
-                    canUseCurrentToolContracts,
-                    hasAssemblyScopedCurrentApplicationUsing);
-                bool canMigrateBareLegacyFirstPartyApi = CanMigrateBareLegacyFirstPartyApiFromUsage(
-                    usage,
-                    canMigrateBareLegacyToolAttribute,
-                    canUseCurrentToolContracts,
-                    hasAssemblyScopedCurrentFirstPartyToolsUsing);
-                bool canUseBareCurrentFirstPartyTools = CanUseBareCurrentFirstPartyToolsFromUsage(
-                    usage,
-                    hasAssemblyScopedCurrentFirstPartyToolsUsing);
-                bool canMigrateBareCurrentDomainContractType =
-                    CanMigrateBareCurrentDomainContractTypeFromUsage(usage, hasAssemblyScopedCurrentDomainUsing);
-                bool hasLocalLegacyMarker = ContainsLegacyToolMigrationMarker(source);
-                bool shouldApplyContractRenames = hasLegacyAssemblySource || hasLocalLegacyMarker;
-                return new CSharpLegacyAssemblyMigrationCapabilities(
-                    canUseCurrentToolContracts,
-                    canUseBareCurrentToolContracts,
-                    CanPreserveBareCurrentToolContractsReferencesFromUsage(
-                        usage,
-                        hasAssemblyScopedCurrentToolContractsUsing),
-                    canMigrateBareLegacyToolAttribute,
-                    canMigrateBareLegacyFirstPartyApi,
-                    canUseBareCurrentFirstPartyTools,
-                    !canUseBareCurrentToolContracts,
-                    canMigrateBareLegacyFirstPartyApi,
-                    canMigrateBareCurrentDomainContractType,
-                    canMigrateBareLegacyApplicationApi,
-                    canMigrateBareLegacyApplicationApi,
-                    CanMigrateAmbiguousBareLegacyToolInfoConstructorFromUsage(
-                        usage,
-                        canMigrateBareLegacyToolAttribute),
-                    shouldApplyContractRenames,
-                    ShouldApplyRegistrarRenamesFromUsage(source, shouldApplyContractRenames),
-                    ShouldApplyDomainMetadataRenamesFromUsage(source, shouldApplyContractRenames));
-            }
-
-            private static CSharpMigrationSourceUsage DetectMigrationSourceUsage(string source)
-            {
-                bool hasCurrentDomainUsingDirective =
-                    RegexMatchesCode(source, CurrentDomainUsingRegex) ||
-                    RegexMatchesCode(source, CurrentDomainGlobalUsingRegex);
-                bool hasCurrentFirstPartyToolsUsingDirective =
-                    RegexMatchesCode(source, CurrentFirstPartyToolsUsingRegex) ||
-                    RegexMatchesCode(source, CurrentFirstPartyToolsGlobalUsingRegex);
-
-                return new CSharpMigrationSourceUsage(
-                    RegexMatchesCode(source, LegacyNamespaceRegex),
-                    RegexMatchesCode(source, LegacyNamespaceUsingRegex),
-                    RegexMatchesCode(source, CurrentApplicationNamespaceRegex),
-                    RegexMatchesCode(source, CurrentDomainNamespaceRegex),
-                    hasCurrentDomainUsingDirective,
-                    RegexMatchesCode(source, CurrentToolContractsNamespaceRegex),
-                    RegexMatchesCode(source, CurrentToolContractsUsingRegex),
-                    RegexMatchesCode(source, CurrentFirstPartyToolsNamespaceRegex),
-                    hasCurrentFirstPartyToolsUsingDirective);
-            }
-
-            private static bool CanUseCurrentToolContractsFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool hasAssemblyScopedCurrentToolContractsUsing)
-            {
-                return usage.HasCurrentToolContractsNamespaceUsage || hasAssemblyScopedCurrentToolContractsUsing;
-            }
-
-            private static bool CanUseBareCurrentToolContractsFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool hasLegacyAssemblySource,
-                bool hasAssemblyScopedCurrentToolContractsUsing)
-            {
-                return hasLegacyAssemblySource ||
-                    usage.HasLegacyNamespaceUsingDirective ||
-                    usage.HasCurrentToolContractsUsingDirective ||
-                    hasAssemblyScopedCurrentToolContractsUsing;
-            }
-
-            private static bool CanPreserveBareCurrentToolContractsReferencesFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool hasAssemblyScopedCurrentToolContractsUsing)
-            {
-                return usage.HasCurrentToolContractsUsingDirective || hasAssemblyScopedCurrentToolContractsUsing;
-            }
-
-            private static bool CanMigrateBareLegacyToolAttributeFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool hasLegacyAssemblySource,
-                string[] legacyNamespaceAliases)
-            {
-                return hasLegacyAssemblySource ||
-                    usage.HasLegacyNamespaceUsage ||
-                    legacyNamespaceAliases.Length > 0;
-            }
-
-            private static bool CanMigrateBareLegacyApplicationApiFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool canMigrateBareLegacyToolAttribute,
-                bool canUseCurrentToolContracts,
-                bool hasAssemblyScopedCurrentApplicationUsing)
-            {
-                return canMigrateBareLegacyToolAttribute ||
-                    canUseCurrentToolContracts ||
-                    hasAssemblyScopedCurrentApplicationUsing ||
-                    usage.HasCurrentApplicationNamespaceUsage;
-            }
-
-            private static bool CanMigrateBareLegacyFirstPartyApiFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool canMigrateBareLegacyToolAttribute,
-                bool canUseCurrentToolContracts,
-                bool hasAssemblyScopedCurrentFirstPartyToolsUsing)
-            {
-                return canMigrateBareLegacyToolAttribute ||
-                    canUseCurrentToolContracts ||
-                    hasAssemblyScopedCurrentFirstPartyToolsUsing ||
-                    usage.HasCurrentFirstPartyToolsNamespaceUsage;
-            }
-
-            private static bool CanUseBareCurrentFirstPartyToolsFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool hasAssemblyScopedCurrentFirstPartyToolsUsing)
-            {
-                return hasAssemblyScopedCurrentFirstPartyToolsUsing ||
-                    usage.HasCurrentFirstPartyToolsUsingDirective;
-            }
-
-            private static bool CanMigrateBareCurrentDomainContractTypeFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool hasAssemblyScopedCurrentDomainUsing)
-            {
-                return hasAssemblyScopedCurrentDomainUsing || usage.HasCurrentDomainUsingDirective;
-            }
-
-            private static bool CanMigrateAmbiguousBareLegacyToolInfoConstructorFromUsage(
-                CSharpMigrationSourceUsage usage,
-                bool canMigrateBareLegacyToolAttribute)
-            {
-                return canMigrateBareLegacyToolAttribute && !usage.HasCurrentDomainNamespaceUsage;
-            }
-
-            private static bool ShouldApplyRegistrarRenamesFromUsage(
-                string source,
-                bool shouldApplyContractRenames)
-            {
-                return shouldApplyContractRenames && RegexMatchesCode(source, LegacyRegistrarRegex);
-            }
-
-            private static bool ShouldApplyDomainMetadataRenamesFromUsage(
-                string source,
-                bool shouldApplyContractRenames)
-            {
-                return shouldApplyContractRenames && RegexMatchesCode(source, LegacyDomainMetadataRegex);
-            }
-
-            private readonly struct CSharpMigrationSourceUsage
-            {
-                public CSharpMigrationSourceUsage(
-                    bool hasLegacyNamespaceUsage,
-                    bool hasLegacyNamespaceUsingDirective,
-                    bool hasCurrentApplicationNamespaceUsage,
-                    bool hasCurrentDomainNamespaceUsage,
-                    bool hasCurrentDomainUsingDirective,
-                    bool hasCurrentToolContractsNamespaceUsage,
-                    bool hasCurrentToolContractsUsingDirective,
-                    bool hasCurrentFirstPartyToolsNamespaceUsage,
-                    bool hasCurrentFirstPartyToolsUsingDirective)
-                {
-                    HasLegacyNamespaceUsage = hasLegacyNamespaceUsage;
-                    HasLegacyNamespaceUsingDirective = hasLegacyNamespaceUsingDirective;
-                    HasCurrentApplicationNamespaceUsage = hasCurrentApplicationNamespaceUsage;
-                    HasCurrentDomainNamespaceUsage = hasCurrentDomainNamespaceUsage;
-                    HasCurrentDomainUsingDirective = hasCurrentDomainUsingDirective;
-                    HasCurrentToolContractsNamespaceUsage = hasCurrentToolContractsNamespaceUsage;
-                    HasCurrentToolContractsUsingDirective = hasCurrentToolContractsUsingDirective;
-                    HasCurrentFirstPartyToolsNamespaceUsage = hasCurrentFirstPartyToolsNamespaceUsage;
-                    HasCurrentFirstPartyToolsUsingDirective = hasCurrentFirstPartyToolsUsingDirective;
-                }
-
-                public bool HasLegacyNamespaceUsage { get; }
-                public bool HasLegacyNamespaceUsingDirective { get; }
-                public bool HasCurrentApplicationNamespaceUsage { get; }
-                public bool HasCurrentDomainNamespaceUsage { get; }
-                public bool HasCurrentDomainUsingDirective { get; }
-                public bool HasCurrentToolContractsNamespaceUsage { get; }
-                public bool HasCurrentToolContractsUsingDirective { get; }
-                public bool HasCurrentFirstPartyToolsNamespaceUsage { get; }
-                public bool HasCurrentFirstPartyToolsUsingDirective { get; }
-            }
         }
     }
 }
