+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationScreenshotCaptureWindowRewriteRules.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19bb34ee4be4431b91b90927a8abab57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
