+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationTypeReplacementRules.cs
@@ -213,188 +213,6 @@ namespace io.github.hatayama.UnityCliLoop.Domain
             return migratedContent;
         }
 
-        public static string ReplaceLegacyApplicationTypeNamesInCode(
-            string source,
-            string[] aliases,
-            string[] currentApplicationNamespaceAliases,
-            bool canMigrateBareLegacyApplicationApi,
-            bool canPreserveBareCurrentToolContractsReferences,
-            string[] assemblyDeclaredTypeNames,
-            ref int replacementCount)
-        {
-            Debug.Assert(source != null, "source must not be null");
-            Debug.Assert(aliases != null, "aliases must not be null");
-            Debug.Assert(
-                currentApplicationNamespaceAliases != null,
-                "currentApplicationNamespaceAliases must not be null");
-            Debug.Assert(assemblyDeclaredTypeNames != null, "assemblyDeclaredTypeNames must not be null");
-
-            string migratedContent = source;
-            foreach (TypeReplacementRule rule in ApplicationTypeReplacementRules)
-            {
-                bool hasProtectedTypeDeclaration = DeclaresLocalType(migratedContent, rule.LegacyName) ||
-                    assemblyDeclaredTypeNames.Contains(rule.LegacyName);
-
-                Regex fullyQualifiedRegex = new(
-                    $@"(?:(?:global::)?{Regex.Escape(LegacyNamespace)}\.){Regex.Escape(rule.LegacyName)}\b",
-                    RegexOptions.Compiled);
-                migratedContent = ReplaceRegexInCode(
-                    migratedContent,
-                    fullyQualifiedRegex,
-                    _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                    ref replacementCount);
-
-                Regex currentApplicationFullyQualifiedRegex = new(
-                    $@"(?:(?:global::)?{Regex.Escape(CurrentApplicationNamespace)}\.){Regex.Escape(rule.CurrentName)}\b",
-                    RegexOptions.Compiled);
-                migratedContent = ReplaceRegexInCode(
-                    migratedContent,
-                    currentApplicationFullyQualifiedRegex,
-                    _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                    ref replacementCount);
-
-                foreach (string alias in aliases)
-                {
-                    Regex aliasRegex = new(
-                        $@"(?<!\w){Regex.Escape(alias)}\.{Regex.Escape(rule.LegacyName)}\b",
-                        RegexOptions.Compiled);
-                    migratedContent = ReplaceRegexInCode(
-                        migratedContent,
-                        aliasRegex,
-                        _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                        ref replacementCount);
-                }
-
-                foreach (string alias in currentApplicationNamespaceAliases)
-                {
-                    Regex currentApplicationAliasRegex = new(
-                        $@"(?<!\w){Regex.Escape(alias)}\.{Regex.Escape(rule.CurrentName)}\b",
-                        RegexOptions.Compiled);
-                    migratedContent = ReplaceRegexInCode(
-                        migratedContent,
-                        currentApplicationAliasRegex,
-                        _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                        ref replacementCount);
-                }
-
-                Regex unqualifiedRegex = new(
-                    $@"(?<![\.:])\b{Regex.Escape(rule.LegacyName)}\b(?!\s*=)",
-                    RegexOptions.Compiled);
-                string unqualifiedReplacement = canPreserveBareCurrentToolContractsReferences
-                    ? rule.CurrentName
-                    : $"{CurrentNamespace}.{rule.CurrentName}";
-                migratedContent = ReplaceRegexInCode(
-                    migratedContent,
-                    unqualifiedRegex,
-                    match => canMigrateBareLegacyApplicationApi &&
-                        !hasProtectedTypeDeclaration &&
-                        ShouldMigrateLegacyTypeReference(migratedContent, rule.LegacyName, match.Index)
-                            ? unqualifiedReplacement
-                            : match.Value,
-                    ref replacementCount);
-            }
-
-            return migratedContent;
-        }
-
-        public static string ReplaceLegacyFirstPartyScreenshotTypeNamesInCode(
-            string source,
-            string[] aliases,
-            string[] currentFirstPartyToolsNamespaceAliases,
-            bool canMigrateBareLegacyFirstPartyScreenshotApi,
-            bool canPreserveBareCurrentToolContractsReferences,
-            string[] assemblyDeclaredTypeNames,
-            ref int replacementCount)
-        {
-            Debug.Assert(source != null, "source must not be null");
-            Debug.Assert(aliases != null, "aliases must not be null");
-            Debug.Assert(
-                currentFirstPartyToolsNamespaceAliases != null,
-                "currentFirstPartyToolsNamespaceAliases must not be null");
-            Debug.Assert(assemblyDeclaredTypeNames != null, "assemblyDeclaredTypeNames must not be null");
-
-            string migratedContent = source;
-            foreach (TypeReplacementRule rule in FirstPartyScreenshotTypeReplacementRules)
-            {
-                bool hasProtectedTypeDeclaration = DeclaresLocalType(migratedContent, rule.LegacyName) ||
-                    assemblyDeclaredTypeNames.Contains(rule.LegacyName);
-
-                Regex fullyQualifiedRegex = new(
-                    $@"(?:(?:global::)?{Regex.Escape(LegacyNamespace)}\.){Regex.Escape(rule.LegacyName)}\b",
-                    RegexOptions.Compiled);
-                migratedContent = ReplaceRegexInCode(
-                    migratedContent,
-                    fullyQualifiedRegex,
-                    _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                    ref replacementCount);
-
-                foreach (string alias in aliases)
-                {
-                    Regex aliasRegex = new(
-                        $@"(?<!\w){Regex.Escape(alias)}\.{Regex.Escape(rule.LegacyName)}\b",
-                        RegexOptions.Compiled);
-                    migratedContent = ReplaceRegexInCode(
-                        migratedContent,
-                        aliasRegex,
-                        _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                        ref replacementCount);
-                }
-
-                if (!string.Equals(
-                        rule.CurrentName,
-                        LegacyEditorWindowCaptureUtilityTypeName,
-                        StringComparison.Ordinal))
-                {
-                    Regex currentFirstPartyFullyQualifiedRegex = new(
-                        $@"(?:(?:global::)?{Regex.Escape(CurrentFirstPartyToolsNamespace)}\.){Regex.Escape(rule.CurrentName)}\b",
-                        RegexOptions.Compiled);
-                    migratedContent = ReplaceRegexInCode(
-                        migratedContent,
-                        currentFirstPartyFullyQualifiedRegex,
-                        _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                        ref replacementCount);
-
-                    foreach (string alias in currentFirstPartyToolsNamespaceAliases)
-                    {
-                        Regex currentFirstPartyAliasRegex = new(
-                            $@"(?<!\w){Regex.Escape(alias)}\.{Regex.Escape(rule.CurrentName)}\b",
-                            RegexOptions.Compiled);
-                        migratedContent = ReplaceRegexInCode(
-                            migratedContent,
-                            currentFirstPartyAliasRegex,
-                            _ => $"{CurrentNamespace}.{rule.CurrentName}",
-                            ref replacementCount);
-                    }
-                }
-
-                Regex unqualifiedRegex = new(
-                    $@"(?<![\.:])\b{Regex.Escape(rule.LegacyName)}\b(?!\s*=)",
-                    RegexOptions.Compiled);
-                if (string.Equals(
-                        rule.CurrentName,
-                        LegacyEditorWindowCaptureUtilityTypeName,
-                        StringComparison.Ordinal))
-                {
-                    continue;
-                }
-
-                string unqualifiedReplacement = canPreserveBareCurrentToolContractsReferences
-                    ? rule.CurrentName
-                    : $"{CurrentNamespace}.{rule.CurrentName}";
-                migratedContent = ReplaceRegexInCode(
-                    migratedContent,
-                    unqualifiedRegex,
-                    match => canMigrateBareLegacyFirstPartyScreenshotApi &&
-                        !hasProtectedTypeDeclaration &&
-                        ShouldMigrateLegacyTypeReference(migratedContent, rule.LegacyName, match.Index)
-                            ? unqualifiedReplacement
-                            : match.Value,
-                    ref replacementCount);
-            }
-
-            return migratedContent;
-        }
-
         public static string ReplaceCurrentDomainContractTypeNamesInCode(
             string source,
             string[] currentDomainNamespaceAliases,
